requirements.txt
requests>=2.28.0
gitpython>=3.1.30
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
azure-devops>=7.1.0
python-sonarqube-api>=1.3.6
retry>=0.9.2
loguru>=0.7.0
langgraph>=0.0.19
pydantic>=2.0.0
streamlit>=1.22.0
pandas>=1.5.3
plotly>=5.14.1
